Backend/requirements.txt
aniso8601==9.0.1
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
Flask==2.2.5
Flask-APScheduler==1.13.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pip==24.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
setuptools==70.0.0
six==1.16.0
SQLAlchemy==1.4.23
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
Werkzeug==2.2.2
zope.interface==6.4.post2
gunicorn==20.0.4
PyMySQL
boto3
Flask-CORS